requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1
psutil==5.9.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
yt-dlp>=2024.12.6,<2026.0.0
celery==5.3.4
redis==5.0.1
flower==2.0.1
pydantic-settings==2.1.0
faster-whisper==1.0.1
torch>=2.0.0
torchaudio>=2.0.0
